code/image-retrieval/requirements.txt
Keras==2.2.4
numpy==1.15.2
pydot==1.2.4
tensorflow-gpu==1.11.0
gensim==3.6.0
pydot==1.2.4
